Updated.
git-svn-id: https://svn.code.sf.net/p/akelpad/codesvn/trunk@1469 24cdac72-59cb-4434-bdd7-86cdcaa6fc6e

+++ akelpad_4/Edit.cpp
@@ -22668,7 +22668,7 @@ BOOL EnsureWindowInMonitor(RECT *rcWindow)
   HMONITOR hMonitor=NULL;
   RECT rcNewWindow;
 
-  if (rcWindow->left == CW_USEDEFAULT)
+  if ((DWORD)rcWindow->left == CW_USEDEFAULT)
     return TRUE;
 
   //Size of the work area on the primary display monitor
